+++ sigle/src/modules/editor/EditorShortcuts/HintsCarousel.tsx
